refactor: 'MoviesCard, Movies, SavedMovies, useMoviesFilterAndLogic, constants': move constant values to constants component
