+++ src/components/landing/Movies/Movies.js
@@ -5,6 +5,10 @@ import Preloader from './Preloader/Preloader';
 import MoviesCardList from '../common/MoviesCardList/MoviesCardList';
 import ButtonElse from './ButtonElse/ButtonElse';
 import { useMoviesFilterAndLogic } from '../../../utils/hooks/useMoviesFilterAndLogic';
+import {
+  getCorrectFormateDuration,
+  getAbsoluteImageUrl,
+} from '../../../utils/constants';
 
 function Movies({
   allMovies,
@@ -26,8 +30,6 @@ function Movies({
     handleInputChange,
     handleSubmit,
     filteredShortMovies,
-    getCorrectFormateDuration,
-    getAbsoluteImageUrl,
     getMovieName,
     setCurrentSearchKeyword,
     setSearchResult,
